requirements.txt
lark-parser==0.9.0
PyYAML==5.3.1
evtx==0.6.11